Add currentMode to Config, refactor Band class to include getCurrentBand method, update frequency handling in Si4735Utils, and adjust display initialization in pico-radio.ino

+++ ScreenSaverDisplay.h
@@ -15,7 +15,7 @@ class ScreenSaverDisplay : public DisplayBase {
     uint8_t saverLineColors[63];  // A vonal színeinek tömbje
 
     SevenSegmentFreq *pSevenSegmentFreq;
-    float currFreq;
+    uint16_t currFreq;
 
    protected:
     /**
